pyproject.toml
[tool.poetry]
name = "localadventuresback"
version = "0.1.0"
description = "Es el back de Local Advetures"
authors = ["JoseJunior <josejuniorpgx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
django = "4.2.7"
django-cors-headers = "4.5.0"
djangorestframework = "3.15.2"
sqlparse = "0.4.4"
typing-extensions = "4.12.2"
poetry = "^1.8.4"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
django-model-utils = "^5.0.0"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
